# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Maratona de Resultados API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{id}/result:
    get:
      operationId: getEventResult
      tags: [results]
      summary: Get event results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployeeEventResult"
...
  /rules/{key}:
    patch:
      operationId: updateRule
      tags: [rules]
      summary: Update rule
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
  /platoon-rules:
    get:
      operationId: getPlatoonRules
      tags: [platoon-rules]
      summary: List platoon rules
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlatoonRule"
    post:
      operationId: createPlatoonRule
      tags: [platoon-rules]
      summary: Create platoon rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatoonRuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatoonRule"
  /platoon-rules/{id}:
    patch:
      operationId: updatePlatoonRule
      tags: [platoon-rules]
      summary: Update platoon rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatoonRuleUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatoonRule"
    delete:
      operationId: deletePlatoonRule
      tags: [platoon-rules]
      summary: Delete platoon rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /results/quarterly:
    get:
      operationId: getQuarterlyResults
      tags: [results]
      summary: Get quarterly results
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: quarter
          in: query
          required: true
          schema:
            type: integer
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: platoon
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuarterlyResult"
  /results/quarterly/close:
    post:
      operationId: closeQuarter
      tags: [results]
      summary: Close a quarter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseQuarterInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuarterCloseResult"
...
components:
  schemas:
...
    Rule:
      type: object
      required: [id, key, value, description]
      properties:
        id:
          type: integer
        key:
          type: string
        value:
          type: string
        description:
          type: string
        updatedAt:
          type: ["string", "null"]
    RuleUpdate:
      type: object
      required: [value]
      properties:
        value:
          type: string
    PlatoonRule:
      type: object
      required: [id, name, color, minScore, maxScore, bonusValue, active, displayOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        minScore:
          type: number
        maxScore:
          type: number
        minInclusive:
          type: boolean
        maxInclusive:
          type: boolean
        bonusValue:
          type: number
        description:
          type: ["string", "null"]
        active:
          type: boolean
        displayOrder:
          type: integer
    PlatoonRuleInput:
      type: object
      required: [name, color, minScore, maxScore, bonusValue]
      properties:
        name:
          type: string
        color:
          type: string
        minScore:
          type: number
        maxScore:
          type: number
        minInclusive:
          type: boolean
        maxInclusive:
          type: boolean
        bonusValue:
          type: number
        description:
          type: string
        displayOrder:
          type: integer
    PlatoonRuleUpdate:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
        minScore:
          type: number
        maxScore:
          type: number
        minInclusive:
          type: boolean
        maxInclusive:
          type: boolean
        bonusValue:
          type: number
        description:
          type: string
        active:
          type: boolean
        displayOrder:
          type: integer
...
    EmployeeEventResult:
      type: object
      required: [employeeId, employeeName, eventId, eventScore]
      properties:
        employeeId:
          type: integer
        employeeName:
          type: string
        eventId:
          type: integer
        eventScore:
          type: number
        projectedPlatoon:
          type: ["string", "null"]
        criteriaDetails:
          type: array
          items:
            $ref: "#/components/schemas/CriterionScoreDetail"
    CriterionScoreDetail:
      type: object
      required: [criterionId, criterionName, normalizedWeight]
      properties:
        criterionId:
          type: integer
        criterionName:
          type: string
        averageScore:
          type: ["number", "null"]
        calibratedScore:
          type: ["number", "null"]
        scoreUsed:
          type: ["number", "null"]
        scorePercentual:
          type: ["number", "null"]
        normalizedWeight:
          type: number
        weightedContribution:
          type: ["number", "null"]
    QuarterlyResult:
      type: object
      required: [employeeId, employeeName, year, quarter, finalResult, platoon, bonusValue]
      properties:
        employeeId:
          type: integer
        employeeName:
          type: string
        year:
          type: integer
        quarter:
          type: integer
        eventsCount:
          type: integer
        grossAverage:
          type: number
        totalAbsences:
          type: integer
        absencePenalty:
          type: number
        finalResult:
          type: number
        platoon:
          type: ["string", "null"]
        platoonColor:
          type: ["string", "null"]
        bonusValue:
          type: number
        eventBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/EventBreakdown"
    EventBreakdown:
      type: object
      required: [eventId, eventName, score]
      properties:
        eventId:
          type: integer
        eventName:
          type: string
        score:
          type: number
    CloseQuarterInput:
      type: object
      required: [year, quarter]
      properties:
        year:
          type: integer
        quarter:
          type: integer
        forced:
          type: boolean
        reason:
          type: string
    QuarterCloseResult:
      type: object
      required: [success, year, quarter, totalProcessed]
      properties:
        success:
          type: boolean
        year:
          type: integer
        quarter:
          type: integer
        totalProcessed:
          type: integer
        warnings:
          type: array
          items:
            type: string
...